+++ .buildkite/pipeline.yml
@@ -173,7 +173,7 @@ steps:
             - src
             - ext
     command: |
-      julia --project --code-coverage=user --color=yes --threads=3 -e '
+      julia --project --code-coverage=user --color=yes --threads=auto -e '
         println("--- :julia: Instantiating project")
         using Pkg
         Pkg.instantiate()
@@ -185,6 +185,7 @@ steps:
         Pkg.activate("docs")
         include("docs/tutorials.jl")
         Pkg.activate("docs")
+        Pkg.instantiate()
         include("docs/make.jl")'
     agents:
       queue: "juliagpu"
