+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Lux"
 uuid = "b2108857-7c20-44ae-9111-449ecde12c47"
 authors = ["Avik Pal <avikpal@mit.edu> and contributors"]
-version = "0.5.16"
+version = "0.5.17"
 
 [deps]
 ADTypes = "47edcb42-4c32-4615-8424-f2b9edc5f35b"
@@ -52,14 +52,16 @@ LuxZygoteExt = "Zygote"
 ADTypes = "0.2"
 Adapt = "4"
 Aqua = "0.8"
+ArrayInterface = "7.7"
 ChainRules = "1.62"
-ChainRulesCore = "1.20"
+ChainRulesCore = "1.21"
 ComponentArrays = "0.15.8"
 ConcreteStructs = "0.2"
 ConstructionBase = "1.5"
 FillArrays = "1"
 Flux = "0.14.11"
 Functors = "0.4.4"
+GPUArraysCore = "0.1.6"
 LinearAlgebra = "1.9"
 Logging = "1.9"
 LuxAMDGPU = "0.2.2"
@@ -70,20 +72,18 @@ LuxLib = "0.3.10"
 LuxTestUtils = "0.1.15"
 MacroTools = "0.5.13"
 Markdown = "1.9"
-Metal = "1"
 Optimisers = "0.3"
 Pkg = "1.9"
 PrecompileTools = "1.2"
 Random = "1.9"
+ReTestItems = "1.23.1"
 Reexport = "1"
 ReverseDiff = "1.15"
-SafeTestsets = "0.1"
 Setfield = "1"
 SparseArrays = "1.9"
 StableRNGs = "1"
 Statistics = "1.9"
 Test = "1.9"
-TestSetExtensions = "3"
 Tracker = "0.2.31"
 TruncatedStacktraces = "1.1"
 WeightInitializers = "0.1.5"
@@ -106,13 +106,13 @@ LuxTestUtils = "ac9de150-d08f-4546-94fb-7472b5760531"
 Optimisers = "3bd65402-5787-11e9-1adc-39752487f4e2"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
+ReTestItems = "817f1d60-ba6b-4fd5-9520-3cf149f6a823"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 Setfield = "efcf1570-3423-57d1-acb7-fd33fddbac46"
 StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-TestSetExtensions = "98d24dd4-01ad-11ea-1b02-c9a08f80db04"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [targets]
-test = ["ADTypes", "Aqua", "ChainRulesCore", "ComponentArrays", "Flux", "Functors", "LuxAMDGPU", "LuxCUDA", "LuxCore", "LuxLib", "LuxTestUtils", "Optimisers", "Random", "SafeTestsets", "Setfield", "StableRNGs", "Statistics", "Test", "TestSetExtensions", "Zygote"]
+test = ["ADTypes", "Aqua", "ChainRulesCore", "ComponentArrays", "Flux", "Functors", "LuxAMDGPU", "LuxCUDA", "LuxCore", "LuxLib", "LuxTestUtils", "Optimisers", "Random", "Setfield", "StableRNGs", "Statistics", "Test", "Zygote", "ReTestItems", "Reexport", "Logging"]
