+++ docs/src/introduction/index.md
@@ -2,7 +2,7 @@
 
 ## Installation
 
-Install [Julia v1.6 or above](https://julialang.org/downloads/). Lux.jl is available through
+Install [Julia v1.9 or above](https://julialang.org/downloads/). Lux.jl is available through
 the Julia package manager. You can enter it by pressing `]` in the REPL and then typing
 
 ```julia
