+++ src/Lux.jl
@@ -13,8 +13,7 @@ PrecompileTools.@recompile_invalidations begin
 
     import LuxCore: AbstractExplicitLayer, AbstractExplicitContainerLayer,
                     initialparameters, initialstates, parameterlength, statelength,
-                    update_state,
-                    trainmode, testmode, setup, apply, display_name
+                    update_state, trainmode, testmode, setup, apply, display_name
     import LuxDeviceUtils: AbstractLuxDevice, AbstractLuxGPUDevice, AbstractLuxDeviceAdaptor
 end
